[INTV] Fixed sprite-to-sprite collision detection at the top and left borders [H@P]
git-svn-id: svn://messdev.no-ip.org/mess@7213 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/video/intv.c
@@ -62,7 +62,7 @@ VIDEO_START( intv )
 
 
 static int sprites_collide(int spriteNum0, int spriteNum1) {
-    UINT8 x1, x2, y1, y2, w1, w2, h1, h2, x0, y0, r0y, r1y,
+    INT16 x1, x2, y1, y2, w1, w2, h1, h2, x0, y0, r0y, r1y,
           width, height, x, y;
 
     struct intv_sprite_type* s1 = &intv_sprite[spriteNum0];
